Fix time on intermission screen
Level time is not displayed correctly on intermission screen. For example,
if it took 100 seconds to complete a level, an engine displays "2:40"
instead of "1:40".

The reason is that "%3.0f" sprintf format rounds its argument (minutes)
to the nearest integer, but we expect its truncation. Well, truncate
it explicitly with 'floor()' routine.

Seconds are rounded too, but it looks more reasonable than truncation.

Reproducible on 'ironwail' quake engine.

Signed-off-by: Kirill K. Smirnov <kirill.k.smirnov@gmail.com>